Migrating functions to services
